+++ auth/src/cleanup/cleanup.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common';
 import { CleanupService } from './cleanup.service';
 import { MongooseModule } from '@nestjs/mongoose';
 import { ScheduleModule } from '@nestjs/schedule';
-import { User, UserSchema } from '../auth/schema/user.shema';
+import { User, UserSchema } from '../modules/auth/schema/user.shema';
 
 @Module({
   imports: [
