+++ auth/src/modules/auth/controller/auth.controller.ts
@@ -29,7 +29,7 @@ import { AssignRoleRequest } from '../dto/request/assign.role.request';
 import { DuplicateResponse } from '../dto/response/dupblicate.response';
 import { RegisterResponse } from '../dto/response/register.response';
 import { LoginResponse } from '../dto/response/login.response';
-import { BasicResponse } from '../../common/responses/basic.response';
+import { BasicResponse } from '../../../common/responses/basic.response';
 import { GetProfileResponse } from '../dto/response/get.profile.response';
 
 /**
