+++ auth/src/modules/auth/dto/response/get.profile.response.ts
@@ -1,5 +1,5 @@
 import { ApiProperty } from '@nestjs/swagger';
-import { BasicResponse } from '../../../common/responses/basic.response';
+import { BasicResponse } from '../../../../common/responses/basic.response';
 
 /**
  * 사용자 프로필 조회 결과를 반환하는 DTO입니다.
