+++ auth/src/saga/saga.module.ts
@@ -4,8 +4,8 @@ import { UserDeletionSagaClient } from './user-deletion/user-deletion.client';
 import { UserDeletionOrchestrator } from './user-deletion/user-deletion.orchestrator';
 import { KafkaModule } from '../kafka/kafka.module';
 import { MongooseModule } from '@nestjs/mongoose';
-import { User, UserSchema } from '../auth/schema/user.shema';
-import { AuthRepository } from '../auth/repository/auth.repository';
+import { User, UserSchema } from '../modules/auth/schema/user.shema';
+import { AuthRepository } from '../modules/auth/repository/auth.repository';
 
 @Module({
   imports: [
